requirements.txt
loguru
psycopg2-binary
sqlalchemy
memory-profiler
python-dotenv